@@ -1,39 +1,39 @@
-module mraft
+module github.com/xkeyideal/mraft
 
-go 1.17
+go 1.21.0
 
 require (
-	github.com/facebookgo/ensure v0.0.0-20160127193407-b4ab57deab51 // indirect
-	github.com/facebookgo/stack v0.0.0-20160209184415-751773369052 // indirect
-	github.com/facebookgo/subset v0.0.0-20150612182917-8dac2c3c4870 // indirect
-	github.com/gin-gonic/gin v1.7.0
-	github.com/google/uuid v1.3.0
-	github.com/lni/dragonboat/v3 v3.3.4
-	github.com/rcrowley/go-metrics v0.0.0-20190826022208-cac0b30c2563
-	github.com/spf13/cobra v1.2.1
-	github.com/tecbot/gorocksdb v0.0.0-20191217155057-f0fad39f321c
+	github.com/cockroachdb/pebble v0.0.0-20210331181633-27fc006b8bfb
+	github.com/gin-gonic/gin v1.9.1
+	github.com/lni/dragonboat/v3 v3.3.7
+	github.com/rcrowley/go-metrics v0.0.0-20201227073835-cf1acfcdf475
+	github.com/spf13/cobra v1.7.0
 	github.com/thrift-iterator/go v0.0.0-20190402154806-9b5a67519118
-	github.com/v2pro/plz v0.0.0-20200805122259-422184e41b6e // indirect
-	github.com/v2pro/quokka v0.0.0-20171201153428-382cb39c6ee6 // indirect
-	github.com/v2pro/wombat v0.0.0-20180402055224-a56dbdcddef2 // indirect
+	github.com/ugorji/go/codec v1.2.11
 	github.com/xkeyideal/gokit v1.4.2
+	go.uber.org/atomic v1.11.0
+	go.uber.org/zap v1.25.0
+	gopkg.in/natefinch/lumberjack.v2 v2.2.1
 )
 
 require (
 	github.com/VictoriaMetrics/metrics v1.6.2 // indirect
 	github.com/armon/go-metrics v0.0.0-20180917152333-f0300d1749da // indirect
-	github.com/cockroachdb/errors v1.7.5 // indirect
+	github.com/bytedance/sonic v1.9.1 // indirect
+	github.com/chenzhuoyu/base64x v0.0.0-20221115062448-fe3a3abad311 // indirect
+	github.com/cockroachdb/errors v1.8.1 // indirect
 	github.com/cockroachdb/logtags v0.0.0-20190617123548-eb05cc24525f // indirect
-	github.com/cockroachdb/pebble v0.0.0-20210331181633-27fc006b8bfb // indirect
-	github.com/cockroachdb/redact v1.0.6 // indirect
+	github.com/cockroachdb/redact v1.0.8 // indirect
 	github.com/cockroachdb/sentry-go v0.6.1-cockroachdb.2 // indirect
+	github.com/gabriel-vasile/mimetype v1.4.2 // indirect
 	github.com/gin-contrib/sse v0.1.0 // indirect
-	github.com/go-playground/locales v0.13.0 // indirect
-	github.com/go-playground/universal-translator v0.17.0 // indirect
-	github.com/go-playground/validator/v10 v10.4.1 // indirect
+	github.com/go-playground/locales v0.14.1 // indirect
+	github.com/go-playground/universal-translator v0.18.1 // indirect
+	github.com/go-playground/validator/v10 v10.14.0 // indirect
+	github.com/goccy/go-json v0.10.2 // indirect
 	github.com/gogo/protobuf v1.3.2 // indirect
 	github.com/golang/protobuf v1.5.2 // indirect
-	github.com/golang/snappy v0.0.3-0.20201103224600-674baa8c7fc3 // indirect
+	github.com/golang/snappy v0.0.4 // indirect
 	github.com/google/btree v1.0.0 // indirect
 	github.com/hashicorp/errwrap v1.0.0 // indirect
 	github.com/hashicorp/go-immutable-radix v1.0.0 // indirect
