@@ -2,11 +2,11 @@ package storage
 
 import (
 	"context"
-	"errors"
+
+	"github.com/xkeyideal/mraft/productready/storage/store"
 
 	"github.com/lni/dragonboat/v3"
 	"github.com/lni/dragonboat/v3/client"
-	"github.com/tecbot/gorocksdb"
 )
 
 type DelCommand struct {
@@ -27,16 +27,17 @@ func (c *DelCommand) RaftInvoke(ctx context.Context, nh *dragonboat.NodeHost, _ 
 	return err
 }
 
-func (c *DelCommand) LocalInvoke(s *Store) error {
-	batch := gorocksdb.NewWriteBatch()
-	defer batch.Destroy()
+func (c *DelCommand) LocalInvoke(s *store.Store, opts ...*WriteOptions) error {
+	batch := s.Batch()
+	defer batch.Close()
 
-	cfHandle, err := s.GetCfHandle(c.CfName)
-	if err != nil {
-		return errors.New("can not find cfName:" + c.CfName)
-	}
+	cf := s.GetColumnFamily(c.CfName)
 
-	batch.DeleteCF(cfHandle, c.Key)
+	batch.Delete(s.BuildColumnFamilyKey(cf, c.Key), s.GetWo())
+
+	// 删除revision
+	revisionKey := buildRevisionKey(c.Key)
+	batch.Delete(s.BuildColumnFamilyKey(cf, revisionKey), s.GetWo())
 
 	return s.Write(batch)
 }
